// src/app/API.d.ts
export type IAPI = IAPIObject & {
    data: {
        updated_artwork?: string;
    };
};